+++ docs/project/changelog.md
@@ -12,6 +12,15 @@ substitutions:
 
 ## [Unreleased]
 
+- {{ API }} {any}`pyodide.runPythonAsync` no longer automatically calls
+  {any}`pyodide.loadPackagesFromImports`.
+  {pr}`1538`.
+- {{ Enhancement }} Added the {any}`PyProxy.callKwargs` method to allow using
+  Python functions with keyword arguments from Javascript.
+  {pr}`1539`
+- {{ Enhancement }} Added the {any}`PyProxy.clone` method.
+  {pr}`1549`
+
 ## Version 0.17.0
 *April 21, 2020*